+++ lang/en/block_course_index.php
@@ -25,4 +25,3 @@
 
 $string['pluginname'] = 'Course Index';
 $string['course_index:addinstance'] = 'Add a new Course Index block';
-$string['course_index:addmyinstance'] = 'Add a new Course Index block';